requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-generativeai
python-dotenv
